% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_race.R
\name{control_race}
\alias{control_race}
\title{Control aspects of the grid search racing process}
\usage{
control_race(
  verbose = FALSE,
  verbose_elim = FALSE,
  allow_par = TRUE,
  extract = NULL,
  save_pred = FALSE,
  burn_in = 3,
  num_ties = 10,
  alpha = 0.05,
  randomize = TRUE,
  pkgs = NULL,
  save_workflow = FALSE,
  event_level = "first",
  parallel_over = "everything"
)
}
\arguments{
\item{verbose_elim}{A logical for whether logging of the elimination of
tuning parameter combinations should occur.}

\item{burn_in}{An integer for how many resamples should be completed for all
grid combinations before parameter filtering begins.}

\item{num_ties}{An integer for when tie-breaking should occur. If there are
two final parameter combinations being evaluated, \code{num_ties} specified how
many more resampling iterations should be evaluated. After \code{num_ties} more
iterations, the parameter combination with the current best results is
retained.}

\item{alpha}{The alpha level for a one-sided confidence interval for each
parameter combination.}

\item{randomize}{Should the resamples be evaluated in a random order?  By
default, the resamples are evaluated in a random order so the random number
seed should be control prior to calling this method (to be reproducible).
For repeated cross-validation the randomization occurs within each repeat.}
}
\value{
An object of class \code{control_race} that echos the argument values.
}
\description{
Control aspects of the grid search racing process
}
\examples{
control_race()
}
