% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_sim_anneal.R
\name{control_sim_anneal}
\alias{control_sim_anneal}
\title{Control aspects of the simulated annealing search process}
\usage{
control_sim_anneal(
  verbose = TRUE,
  no_improve = Inf,
  restart = 8L,
  radius = c(0.05, 0.15),
  flip = 3/4,
  cooling_coef = 0.02,
  extract = NULL,
  save_pred = FALSE,
  time_limit = NA,
  pkgs = NULL,
  save_workflow = FALSE,
  save_history = FALSE,
  event_level = "first",
  parallel_over = NULL,
  allow_par = TRUE
)
}
\arguments{
\item{no_improve}{The integer cutoff for the number of iterations without
better results.}

\item{restart}{The number of iterations with no improvement before new tuning
parameter candidates are generated from the last, overall best conditions.}

\item{radius}{Two real numbers on \verb{(0, 1)} describing what a value "in the
neighborhood" of the current result should be. If all numeric parameters were
scaled to be on the \verb{[0, 1]} scale, these values set the min. and max.
of a radius of a circle used to generate new numeric parameter values.}

\item{flip}{A real number between \verb{[0, 1]} for the probability of changing
any non-numeric parameter values at each iteration.}

\item{cooling_coef}{A real, positive number to influence the cooling
schedule. Larger values decrease the probability of accepting a sub-optimal
parameter setting.}

\item{time_limit}{A number for the minimum number of \emph{minutes} (elapsed) that
the function should execute. The elapsed time is evaluated at internal
checkpoints and, if over time, the results at that time are returned (with
a warning). This means that the \code{time_limit} is not an exact limit, but a
minimum time limit.}

\item{save_history}{A logical to save the iteration details of the search.
These are saved to \code{tempdir()} named \code{sa_history.RData}. These results are
deleted when the R session ends. This option is only useful for teaching
purposes.}
}
\value{
An object of class \code{control_sim_anneal} that echos the argument values.
}
\description{
Control aspects of the simulated annealing search process
}
\examples{
control_sim_anneal()
}
