% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finetune-package.R
\docType{package}
\name{finetune-package}
\alias{finetune}
\alias{finetune-package}
\title{finetune: Additional Functions for Model Tuning}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The ability to tune models is important. 'finetune' enhances the 'tune' package by providing more specialized methods for finding reasonable values of model tuning parameters. Two racing methods described by Kuhn (2014) \doi{10.48550/arXiv.1405.6974} are included. An iterative search method using generalized simulated annealing (Bohachevsky, Johnson and Stein, 1986) \doi{10.1080/00401706.1986.10488128} is also included.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tidymodels/finetune}
  \item \url{https://finetune.tidymodels.org}
  \item Report bugs at \url{https://github.com/tidymodels/finetune/issues}
}

}
\author{
\strong{Maintainer}: Max Kuhn \email{max@posit.co} (\href{https://orcid.org/0000-0003-2402-136X}{ORCID})

Other contributors:
\itemize{
  \item Posit Software, PBC (03wc8by49) [copyright holder, funder]
}

}
\keyword{internal}
