% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_race.R
\name{plot_race}
\alias{plot_race}
\title{Plot racing results}
\usage{
plot_race(x)
}
\arguments{
\item{x}{A object with class \code{tune_results}}
}
\value{
A ggplot object.
}
\description{
Plot difference parameter combinations versus the number of times that each
was evaluated. The plot colors the results by the performanc metric that was
being optimized/tested.
}
