% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_nested.R
\name{tune_nested}
\alias{tune_nested}
\alias{tune_nested.model_spec}
\alias{tune_nested.workflow}
\title{Nested resampling for estimating tuning performance}
\usage{
tune_nested(object, ...)

\method{tune_nested}{model_spec}(
  object,
  preprocessor,
  resamples,
  ...,
  fn = "tune_grid",
  nest_control = control_nested()
)

\method{tune_nested}{workflow}(
  object,
  resamples,
  ...,
  fn = "tune_grid",
  nest_control = control_nested()
)
}
\arguments{
\item{object}{A \code{parsnip} model specification or an unfitted
\link[workflows:workflow]{workflow()}. No tuning parameters are allowed; if arguments
have been marked with \link[hardhat:tune]{tune()}, their values must be
\link[tune:finalize_model]{finalized}.}

\item{...}{Other options to pass to the tuning function (e.g., \code{metrics} etc).}

\item{preprocessor}{A traditional model formula or a recipe created using
\code{\link[recipes:recipe]{recipes::recipe()}}.}

\item{resamples}{An object produced by \code{\link[rsample:nested_cv]{rsample::nested_cv()}}.}

\item{fn}{A character string for the procedure used to tune the model.
Possible values are: \code{"tune_bayes"}, \code{"tune_grid"}, \code{"tune_race_anova"},
\code{"tune_race_win_loss"}, or \code{"tune_sim_anneal"}.}

\item{nest_control}{An object used to modify the tuning process.}
}
\value{
An object of class \code{"nested_results"}.
}
\description{
Using the same model tuning results to find optimal settings \emph{and} to measure
optimal performance might lead to optimistic performance estimates.
\code{\link[=tune_nested]{tune_nested()}} resamples the entire tuning process and uses out-of-sample
data to compute performance statistics.
}
\seealso{
\url{tune_bayes}, \url{tune_grid}, \url{tune_race_anova},
\url{tune_race_win_loss}, \url{tune_sim_anneal}
}
