% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_sim_anneal.R
\name{tune_sim_anneal}
\alias{tune_sim_anneal}
\alias{tune_sim_anneal.model_spec}
\alias{tune_sim_anneal.workflow}
\title{Optimization of model parameters via simulated annealing}
\usage{
tune_sim_anneal(object, ...)

\method{tune_sim_anneal}{model_spec}(
  object,
  preprocessor,
  resamples,
  ...,
  iter = 10,
  param_info = NULL,
  metrics = NULL,
  objective = NULL,
  initial = 5,
  control = control_sim_anneal()
)

\method{tune_sim_anneal}{workflow}(
  object,
  resamples,
  ...,
  iter = 10,
  param_info = NULL,
  metrics = NULL,
  objective = NULL,
  initial = 5,
  control = control_sim_anneal()
)
}
\arguments{
\item{...}{Not currently used.}

\item{iter}{The maximum number of search iterations.}

\item{metrics}{A \code{\link[yardstick:metric_set]{yardstick::metric_set()}} object containing information on how
models will be evaluated for performance. The first metric in \code{metrics} is the
one that will be optimized.}

\item{objective}{A character string for what metric should be optimized or
an acquisition function object.}

\item{initial}{An initial set of results in a tidy format (as would result
from \code{\link[=tune_grid]{tune_grid()}}) or a positive integer. It is suggested that the number of
initial results be greater than the number of parameters being optimized.}

\item{control}{A control object created by \code{\link[=control_sim_anneal]{control_sim_anneal()}}}
}
\value{
A tibble of results that mirror those generated by \code{\link[=tune_grid]{tune_grid()}}.
However, these results contain an \code{.iter} column and replicate the \code{rset}
object multiple times over iterations (at limited additional memory costs).
}
\description{
\code{\link[=tune_sim_anneal]{tune_sim_anneal()}} uses models to generate new candidate tuning parameter
combinations based on previous results.
}
