% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_sim_anneal.R
\name{control_sim_anneal}
\alias{control_sim_anneal}
\title{Control aspects of the simulated annealing search process}
\usage{
control_sim_anneal(
  verbose = FALSE,
  no_improve = Inf,
  restart = 8L,
  radius = 0.1,
  flip = 0.2,
  cooling_coef = 0.02,
  extract = NULL,
  save_pred = FALSE,
  time_limit = NA,
  pkgs = NULL,
  save_workflow = FALSE
)
}
\arguments{
\item{verbose}{A logical for logging results as they are generated. Despite
this argument, warnings and errors are always shown. If using a dark IDE
theme, some logging messages might be hard to see. If this is the case,
try setting the \code{tidymodels.dark} option with
\code{options(tidymodels.dark = TRUE)} to print lighter colors.}

\item{no_improve}{The integer cutoff for the number of iterations without
better results.}

\item{restart}{The number of iterations with no improvement before new tuning
parameter candidates are generated from the last, overall best conditions.}

\item{radius}{A real number on \verb{(0, 1)} describing what a value "in the
neighborhood" of the current result should be. If all numeric parameters were
scaled to be on the \verb{[0, 1]} scale, this is the radius of a circle used to
generate new numeric parameter values.}

\item{flip}{A real number between \verb{[0, 1]} for the probability of changing
any non-numeric parameter values at each iteration.}

\item{cooling_coef}{A real, positive number to influence the cooling
schedule. Larger values decrease the probability of accepting a sub-optimal
parameter setting.}

\item{extract}{An optional function with at least one argument (or \code{NULL})
that can be used to retain arbitrary objects from the model fit object,
recipe, or other elements of the workflow.}

\item{save_pred}{A logical for whether the out-of-sample predictions should
be saved for each model \emph{evaluated}.}

\item{time_limit}{A number for the minimum number of \emph{minutes} (elapsed) that
the function should execute. The elapsed time is evaluated at internal
checkpoints and, if over time, the results at that time are returned (with
a warning). This means that the \code{time_limit} is not an exact limit, but a
minimum time limit.}

\item{pkgs}{An optional character string of R package names that should be
loaded (by namespace) during parallel processing.}

\item{save_workflow}{A logical for whether the workflow should be appended
to the output as an attribute.}
}
\description{
Control aspects of the simulated annealing search process
}
