% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_race.R
\name{plot_race}
\alias{plot_race}
\title{Plot racing results}
\usage{
plot_race(x)
}
\arguments{
\item{x}{A object with class \code{tune_results}}
}
\value{
A ggplot object.
}
\description{
Plot the model results over stages of the racing results. A line is given
for each submodel that was tested.
}
